+++ packages/core/src/portal/Popover.tsx
@@ -14,8 +14,10 @@ import PopoverBody from './PopoverBody';
 import PopoverOverlay from './PopoverOverlay';
 import Portal, { RenderContent, RenderTarget } from './Portal';
 import Position from './Position';
+import uniqueId from '../utils/uniqueId';
 
 interface PopoverProps {
+    ariaControlsSuffix?: string;
     closeOnClickOutside: boolean;
     closeOnEscape: boolean;
     content: RenderContent;
@@ -24,17 +26,34 @@ interface PopoverProps {
     target: RenderTarget;
 }
 
-const Popover: React.FC<PopoverProps> = ({ closeOnClickOutside, closeOnEscape, content, offset, position, target }) => {
+const Popover: React.FC<PopoverProps> = ({
+    ariaControlsSuffix,
+    closeOnClickOutside,
+    closeOnEscape,
+    content,
+    offset,
+    position,
+    target,
+}) => {
     const targetRef = React.useRef<HTMLDivElement>();
+    const controlsSuffix = ariaControlsSuffix || `${uniqueId()}`;
 
     const renderTarget = (toggle: Toggle, opened: boolean): React.ReactElement => (
-        <div ref={targetRef}>{target(toggle, opened)}</div>
+        <div
+            ref={targetRef}
+            aria-expanded={opened ? 'true' : 'false'}
+            aria-haspopup="dialog"
+            aria-controls={`rpv-core__popver-body-${controlsSuffix}`}
+        >
+            {target(toggle, opened)}
+        </div>
     );
 
     const renderContent = (toggle: Toggle): React.ReactElement => (
         <>
             <PopoverOverlay closeOnEscape={closeOnEscape} onClose={toggle} />
             <PopoverBody
+                ariaControlsSuffix={controlsSuffix}
                 closeOnClickOutside={closeOnClickOutside}
                 offset={offset}
                 position={position}