+++ packages/get-file/src/DownloadButton.tsx
@@ -22,7 +22,7 @@ const DownloadButton: React.FC<RenderDownloadProps> = ({ onClick }) => {
         <Tooltip
             position={Position.BottomCenter}
             target={
-                <MinimalButton label={label as string} onClick={onClick}>
+                <MinimalButton ariaLabel={label as string} onClick={onClick}>
                     <DownloadIcon />
                 </MinimalButton>
             }