+++ packages/page-navigation/src/CurrentPageInput.tsx
@@ -78,7 +78,7 @@ const CurrentPageInput: React.FC<{
     return (
         <span className="rpv-page-navigation__current-page-input">
             <TextBox
-                label={label as string}
+                ariaLabel={label as string}
                 type="text"
                 value={editingPage}
                 onChange={setEditingPage}