+++ packages/properties/src/ShowPropertiesButton.tsx
@@ -22,7 +22,7 @@ const ShowPropertiesButton: React.FC<RenderShowPropertiesProps> = ({ onClick }) 
         <Tooltip
             position={Position.BottomCenter}
             target={
-                <MinimalButton label={label as string} onClick={onClick}>
+                <MinimalButton ariaLabel={label as string} onClick={onClick}>
                     <InfoIcon />
                 </MinimalButton>
             }