+++ packages/rotate/src/RotateButton.tsx
@@ -28,7 +28,7 @@ const RotateButton: React.FC<RenderRotateProps> = ({ direction, onClick }) => {
         <Tooltip
             position={Position.BottomCenter}
             target={
-                <MinimalButton label={label as string} onClick={onClick}>
+                <MinimalButton ariaLabel={label as string} onClick={onClick}>
                     {icon}
                 </MinimalButton>
             }